+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.4]
+- added volumes to Data (GOOG.csv to be replaced for compare example)
+- added max drawdown and sharpe ratio lo report
+- added new strategy modifier skip_first if you want to avoid entering as soon as a signal is available
+
 ## [0.1.3]
 - Possibility to display the list of trades produced by a strategy
 - Minor bugs fixes